refactor: split off value casting out of config:system:set command
