+++ core/Command/Config/System/SetConfig.php
@@ -17,6 +17,7 @@ use Symfony\Component\Console\Output\OutputInterface;
 class SetConfig extends Base {
 	public function __construct(
 		SystemConfig $systemConfig,
+		private CastHelper $castHelper,
 	) {
 		parent::__construct($systemConfig);
 	}
@@ -57,7 +58,7 @@ class SetConfig extends Base {
 	protected function execute(InputInterface $input, OutputInterface $output): int {
 		$configNames = $input->getArgument('name');
 		$configName = $configNames[0];
-		$configValue = $this->castValue($input->getOption('value'), $input->getOption('type'));
+		$configValue = $this->castHelper->castValue($input->getOption('value'), $input->getOption('type'));
 		$updateOnly = $input->getOption('update-only');
 
 		if (count($configNames) > 1) {
@@ -78,80 +79,6 @@ class SetConfig extends Base {
 
 		$output->writeln('<info>System config value ' . implode(' => ', $configNames) . ' set to ' . $configValue['readable-value'] . '</info>');
 		return 0;
-	}
-
-	/**
-	 * @param string $value
-	 * @param string $type
-	 * @return mixed
-	 * @throws \InvalidArgumentException
-	 */
-	protected function castValue($value, $type) {
-		switch ($type) {
-			case 'integer':
-			case 'int':
-				if (!is_numeric($value)) {
-					throw new \InvalidArgumentException('Non-numeric value specified');
-				}
-				return [
-					'value' => (int)$value,
-					'readable-value' => 'integer ' . (int)$value,
-				];
-
-			case 'double':
-			case 'float':
-				if (!is_numeric($value)) {
-					throw new \InvalidArgumentException('Non-numeric value specified');
-				}
-				return [
-					'value' => (float)$value,
-					'readable-value' => 'double ' . (float)$value,
-				];
-
-			case 'boolean':
-			case 'bool':
-				$value = strtolower($value);
-				switch ($value) {
-					case 'true':
-						return [
-							'value' => true,
-							'readable-value' => 'boolean ' . $value,
-						];
-
-					case 'false':
-						return [
-							'value' => false,
-							'readable-value' => 'boolean ' . $value,
-						];
-
-					default:
-						throw new \InvalidArgumentException('Unable to parse value as boolean');
-				}
-
-				// no break
-			case 'null':
-				return [
-					'value' => null,
-					'readable-value' => 'null',
-				];
-
-			case 'string':
-				$value = (string)$value;
-				return [
-					'value' => $value,
-					'readable-value' => ($value === '') ? 'empty string' : 'string ' . $value,
-				];
-
-			case 'json':
-				$value = json_decode($value, true);
-				return [
-					'value' => $value,
-					'readable-value' => 'json ' . json_encode($value),
-				];
-
-			default:
-				throw new \InvalidArgumentException('Invalid type');
-		}
 	}
 
 	/**
